Move background job registration to install/update
