+++ appinfo/info.xml
@@ -15,6 +15,7 @@
 	<licence>AGPL</licence>
 	<author>Frank Karlitschek, Joas Schilling</author>
 	<shipped>true</shipped>
+	<version>2.1.3</version>
 	<default_enable/>
 	<types>
 		<filesystem/>
