+++ appinfo/install.php
@@ -1,0 +1,5 @@
+<?php
+
+// Cron job for sending emails and pruning the activity list
+\OC::$server->getJobList()->add('OCA\Activity\BackgroundJob\EmailNotification');
+\OC::$server->getJobList()->add('OCA\Activity\BackgroundJob\ExpireActivities');
