+++ tests/appinfo/apptest.php
@@ -20,7 +20,9 @@
  * License along with this library.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-namespace OCA\Activity\Tests;
+namespace OCA\Activity\Tests\AppInfo;
+
+use OCA\Activity\Tests\TestCase;
 
 class AppTest extends TestCase {
 	public function testNavigationEntry() {
@@ -32,16 +34,6 @@ class AppTest extends TestCase {
 
 		// Test whether the navigation entry got added
 		$this->assertCount(1, $navigationManager->getAll());
-	}
-
-	public function testJobList() {
-		$jobList = \OC::$server->getJobList();
-
-		require '../appinfo/app.php';
-
-		// Test whether the background jobs got registered
-		$this->assertTrue($jobList->has('OCA\Activity\BackgroundJob\EmailNotification', null));
-		$this->assertTrue($jobList->has('OCA\Activity\BackgroundJob\ExpireActivities', null));
 	}
 
 // FIXME: Uncomment once the OC_App stuff is not static anymore
